worked on combat system
created a new branch to work on the combat system. been working on a system to have multiple enemies show up on the game map. Still working on it. added a player level property to the player object, this determines what enemies show up on the game map. If i set it to 1, the enemy icon takes me to the goblin fight, if i set it to 2, it takes me to the Orc fight. I need to find a way to have both buttons on the game map, letting me to choose either goblin or orc from there.